+++ app/resources/numismatics/issues/numismatics/issue.rb
@@ -62,6 +62,7 @@ module Numismatics
     attribute :visibility
     attribute :workflow_note, Valkyrie::Types::Array.of(WorkflowNote).optional
     attribute :pending_uploads, Valkyrie::Types::Array.of(PendingUpload)
+    attribute :claimed_by, Valkyrie::Types::String
 
     # manifest metadata
     attribute :start_canvas