+++ app/resources/raster_resources/raster_resource_change_set.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 class RasterResourceChangeSet < ChangeSet
   apply_workflow(GeoWorkflow)
+  enable_claiming
   delegate :human_readable_type, to: :model
 
   include GeoChangeSetProperties