+++ app/resources/vector_resources/vector_resource_change_set.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 class VectorResourceChangeSet < ChangeSet
   apply_workflow(GeoWorkflow)
+  enable_claiming
   delegate :human_readable_type, to: :model
 
   include GeoChangeSetProperties